refactor routes and handlers
