+++ lib/index.js
@@ -17,7 +17,9 @@ Glue.compose(manifest, options, (err, server) => {
 
   // Step 2.
   // Show the server to our instance of labbable
-  labbable.using(server)
+  if (process.env.NODE_ENV === 'test') {
+    labbable.using(server)
+  }
 
   // Step 3.
   // Initialize your server
