+++ lib/modules/api/profiles/index.js
@@ -1,117 +1,7 @@
-const inputValidations = require('./validations/input')
-const outputValidations = require('./validations/output')
-const replyHelper = require('../helpers')
+const Routes = require('./routes')
 
 const register = (server, options, next) => {
-  // server.dependency('auth')
-  server.route([
-    // Get /profiles/{username}
-    {
-      method: 'GET',
-      path: '/profiles/{username}',
-      config: {
-        auth: {mode: 'optional', strategy: 'jwt'},
-        response: outputValidations.ProfileOnGetOutputValidationConfig,
-        validate: inputValidations.ProfileParamsValidation,
-        description: 'Get the  user profile',
-        tags: ['api', 'profiles']
-      },
-      handler: (request, reply) => {
-        var username = request.params.username
-        server.methods.services.users.getByUsername(username, (err, user) => {
-          if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-
-          if (!user) {
-            return reply({
-              errors: {
-                404: [`User with username "${username} not found !"`]
-              }
-            }).code(422)
-          }
-
-          if (request.auth.isAuthenticated) {
-            return reply(constructProfileResponse(user, request.auth.credentials.user))
-          }
-
-          return reply(constructProfileResponse(user))
-        })
-      }
-    },
-    // POST /profiles/{username}/follow
-    {
-      method: 'POST',
-      path: '/profiles/{username}/follow',
-      config: {
-        auth: 'jwt',
-        response: outputValidations.ProfileFollowOutputValidationConfig,
-        validate: inputValidations.ProfileParamsValidation,
-        description: 'Follow/Unfollow a user',
-        tags: ['api', 'profiles'],
-        notes: '__Require a valid JWT Token__'
-      },
-      handler: (request, reply) => {
-        let username = request.params.username
-        let currentUser = request.auth.credentials.user
-
-        server.methods.services.users.getByUsername(username, (err, userToFollow) => {
-          if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-
-          if (!userToFollow) {
-            return reply({
-              errors: {
-                404: [`User with username "${username} not found !"`]
-              }
-            }).code(404)
-          }
-
-          server.methods.services.users.follow(currentUser, userToFollow._id, (err, me) => {
-            if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-            return reply(constructProfileResponse(userToFollow, me))
-          })
-        })
-      }
-    },
-    // DELETE /profiles/{username}/follow
-    {
-      method: 'DELETE',
-      path: '/profiles/{username}/follow',
-      config: {
-        auth: 'jwt',
-        response: outputValidations.ProfileFollowOutputValidationConfig,
-        validate: inputValidations.ProfileParamsValidation,
-        description: 'Follow/Unfollow a user',
-        tags: ['api', 'profiles'],
-        notes: '__Require a valid JWT Token__'
-      },
-      handler: (request, reply) => {
-        let username = request.params.username
-        let currentUser = request.auth.credentials.user
-
-        server.methods.services.users.getByUsername(username, (err, userToFollow) => {
-          if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-
-          if (!userToFollow) {
-            return reply({
-              errors: {
-                404: [`User with username "${username} not found !"`]
-              }
-            }).code(404)
-          }
-
-          server.methods.services.users.unfollow(currentUser, userToFollow._id, (err, me) => {
-            if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-            return reply(constructProfileResponse(userToFollow, me))
-          })
-        })
-      }
-    }
-  ])
-
-  function constructProfileResponse (user, authenticatedUser = null) {
-    let profile = { user: user.toProfileJSONFor(authenticatedUser) }
-    return profile
-  }
-
+  server.route(Routes(server))
   return next()
 }
