+++ lib/modules/api/profiles/package.json
@@ -1,4 +1,4 @@
 {
     "name": "profiles",
-    "versopm": "1.0.0"
+    "version": "1.0.0"
 }
