+++ lib/modules/api/tags/index.js
@@ -1,26 +1,7 @@
-const outputValidations = require('./validations/output')
-const replyHelper = require('../helpers')
+const Routes = require('./routes')
 
 const register = (server, options, next) => {
-  server.route([
-    // GET /api/tags
-    {
-      method: 'GET',
-      path: '/tags',
-      config: {
-        response: outputValidations.TagListOutputValidationsConfig,
-        description: 'Get all tags',
-        tags: ['api', 'tags']
-      },
-      handler: (request, reply) => {
-        server.methods.services.tags.getAll((err, tags) => {
-          if (err) return reply(replyHelper.constructErrorResponse(err)).code(422)
-          return reply({tags})
-        })
-      }
-    }
-  ])
-
+  server.route(Routes(server))
   return next()
 }
