+++ lib/modules/models/index.js
@@ -5,7 +5,6 @@ const bluebird = require('bluebird')
 const register = (server, options, next) => {
   mongoose.Promise = bluebird
 
-  // TODO: Uppgrade to mongoose@4.11.0
   mongoose.connect(config.database.uri, config.database.options, (err, db) => {
     if (err) console.log(err)
